refactor: refactor common functions
