+++ pitr/cli/internal/cmd/common.go
@@ -19,10 +19,13 @@ package cmd
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/apache/shardingsphere-on-cloud/pitr/cli/internal/pkg"
 	"github.com/apache/shardingsphere-on-cloud/pitr/cli/internal/pkg/model"
 	"github.com/apache/shardingsphere-on-cloud/pitr/cli/internal/pkg/xerr"
+	"github.com/apache/shardingsphere-on-cloud/pitr/cli/pkg/logging"
+	"github.com/jedib0t/go-pretty/v6/table"
 )
 
 func validate(ls pkg.ILocalStorage, csn, recordID string) ([]*model.LsBackup, error) {
@@ -53,4 +56,106 @@ func validate(ls pkg.ILocalStorage, csn, recordID string) ([]*model.LsBackup, er
 	}
 
 	return baks, nil
+}
+
+func convertLocalhost(ip string) string {
+	if ip == "127.0.0.1" {
+		return Host
+	}
+	return ip
+}
+
+func checkAgentServerStatus(lsBackup *model.LsBackup) bool {
+
+	statusList := make([]*model.AgentServerStatus, 0)
+
+	// all agent server are available
+	available := true
+
+	// IMPORTANT: we don't support multiple agent server run on the same host
+	asMap := make(map[string]bool)
+	asDuplicate := false
+
+	for _, node := range lsBackup.SsBackup.StorageNodes {
+		sn := node
+		as := pkg.NewAgentServer(fmt.Sprintf("%s:%d", convertLocalhost(sn.IP), AgentPort))
+		in := &model.HealthCheckIn{
+			DBPort:   sn.Port,
+			DBName:   sn.Database,
+			Username: sn.Username,
+			Password: sn.Password,
+		}
+		if err := as.CheckStatus(in); err != nil {
+			statusList = append(statusList, &model.AgentServerStatus{IP: sn.IP, Port: AgentPort, Status: fmt.Sprintf("Unavailable: %s", err)})
+			available = false
+		} else {
+			statusList = append(statusList, &model.AgentServerStatus{IP: sn.IP, Port: AgentPort, Status: "Available"})
+		}
+	}
+
+	t := table.NewWriter()
+	t.SetOutputMirror(os.Stdout)
+	t.SetTitle("Agent Server Status")
+	t.AppendHeader(table.Row{"#", "Agent Server IP", "Agent Server Port", "Status"})
+
+	for i, s := range statusList {
+		t.AppendRow([]interface{}{i + 1, s.IP, s.Port, s.Status})
+		t.AppendSeparator()
+	}
+
+	t.Render()
+
+	for _, node := range lsBackup.SsBackup.StorageNodes {
+		if _, ok := asMap[node.IP]; ok {
+			asDuplicate = true
+			break
+		}
+		asMap[node.IP] = true
+	}
+
+	if asDuplicate {
+		logging.Error("IMPORTANT!: we don't support multiple agent server run on the same host.\n")
+		return false
+	}
+
+	return available
+}
+
+func checkDiskSpace(lsBackup *model.LsBackup) error {
+	var (
+		diskspaceList = make([]*model.DiskSpaceStatus, 0)
+	)
+	for _, sn := range lsBackup.SsBackup.StorageNodes {
+		var data string
+		as := pkg.NewAgentServer(fmt.Sprintf("%s:%d", convertLocalhost(sn.IP), AgentPort))
+		in := &model.DiskSpaceIn{
+			DiskPath: BackupPath,
+		}
+
+		out, err := as.ShowDiskSpace(in)
+
+		if err != nil {
+			data = "Check disk space failed."
+		} else {
+			data = out.Data
+		}
+
+		diskspaceList = append(diskspaceList, &model.DiskSpaceStatus{
+			IP:              sn.IP,
+			Path:            BackupPath,
+			DiskSpaceStatus: data,
+		})
+	}
+
+	// print diskspace result formatted
+	t := table.NewWriter()
+	t.SetOutputMirror(os.Stdout)
+	t.SetTitle("Disk Space Status")
+	t.AppendHeader(table.Row{"#", "Data Node IP", "Disk Path", "Disk Space Status"})
+	for i, ds := range diskspaceList {
+		t.AppendRow([]interface{}{i + 1, ds.IP, ds.Path, ds.DiskSpaceStatus})
+		t.AppendSeparator()
+	}
+	t.Render()
+	return nil
 }
